+++ src/GraphicEngine/Image.h
@@ -24,8 +24,8 @@ typedef struct ANIMATION
 {
     sfSprite *sprite;           // Image convertie en Sprite de l'animation
     //Coordonnée de la premiere case
-    float x;
-    float y;
+    int x;
+    int y;
     //Taille d'une case
     int image_hauteur;
     int image_largeur;
@@ -33,13 +33,13 @@ typedef struct ANIMATION
     int nombre_image;
     int cur_image;              // Position dans l'animation
     int play;                   // Nombre de lectures à effectuer
-    int fps;                    // Temps d'attente entre chaque image (en ms)
+    float fps;                  // Temps d'attente entre chaque image (en ms)
 
     sfClock *clock;             // Timer
 } Animation;
 
 //Animation* animation_Create();
-Animation* animation_Create(sfImage*, float, float, int, int, int, int, int, float);
+Animation* animation_Create(sfImage*, int, int, int, int, int, int, int, float);
 void animation_Destroy(Animation*);
 
 void animation_Play(Animation*, int);
